// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/lesson-page-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { LessonSummaryHttpResponse } from './lesson-summary-http-response';

export interface LessonPageHttpResponse {
    availableCategories?: Array<string>;
    items?: Array<LessonSummaryHttpResponse>;
    page?: number;
    pageSize?: number;
    readLessonCount?: number;
    totalCount?: number;
    totalLessonCount?: number;
    totalPages?: number;
}