// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/lesson-summary-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface LessonSummaryHttpResponse {
    category?: string;
    difficulty?: string;
    estimatedReadMinutes?: number;
    id?: string;
    isRead?: boolean;
    summary?: string | null;
    title?: string;
}